Add files via upload
Introduction
Through this project work, we aim to create a simplified version of a banking application.
The application will act as a portal which provides it’s users with various banking facilities
like account management and transaction management. The aim is to learn how to create
an application which uses a database in its backend. We also plan to make the application
safe and secure so that it is not susceptible to vulnerabilities and security attacks.
Approach
We have divided the project work in the following stages:
1. We will first decide the tables and the respective attributes needed for our database.
2. We will then sketch an ER diagram of the database.
3. We will then plan out the appropriate screens needed for our application.
4. We will then move on to the programming part of the application.
5. We will then create the associated tests for our application to ensure that it works
correctly.
Hardware Specification
● PC with 2 GB hard-disk and 256 MB RAM
Software Specification
● Xampp - 7.1.1
● MySQL
● Sublime Text Editor